requirements.txt
simpleitk~=2.0.2
scikit-learn~=1.3.2
cachetools~=4.2.4
numpy~=1.23.5
pandas~=2.0.3
scipy~=1.9.3
tabulate~=0.8.9
matplotlib~=3.3.4
scikit-image~=0.19.3
tensorflow~=2.5.0
joblib~=1.4.2
dill~=0.3.8
pygments~=2.18.0
psutil~=5.8.0
wandb~=0.17.0
seaborn~=0.11.2
h5py~=3.1.0
itk~=5.4.0
voluptuous~=0.12.2
humanfriendly~=10.0
click~=8.1.3
natsort~=7.1.1
watchdog~=2.1.9
colorama~=0.4.4
coloredlogs~=15.0.1
pillow~=10.3.0
pdf2dcm~=0.5.0
pydicom~=2.4.4
streamlit~=1.35.0